+++ packages/werkbank/lib/src/_internal/src/widgets/src/panels/configuration_panel/tab_addon_control_section_list.dart
@@ -65,14 +65,14 @@ class _AddonControlSectionList extends StatefulWidget {
 
 class _AddonControlSectionListState extends State<_AddonControlSectionList> {
   List<_AddonControlSectionWithAddonId>? _orderedSections;
-  late PanelTabsController _panelTabsController;
+  late SectionsController _sectionsController;
 
   void _updateSections() {
     final orderedList = widget.addonControlSections.toList()
       ..sort((a, b) => a.sortHint.compareTo(b.sortHint));
 
     setState(() {
-      _orderedSections = _panelTabsController.addAndOrderSections(
+      _orderedSections = _sectionsController.addAndOrderSections(
         tab: widget.tab,
         sections: orderedList,
         getId: (section) => section.id,
@@ -83,7 +83,7 @@ class _AddonControlSectionListState extends State<_AddonControlSectionList> {
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
-    _panelTabsController = GlobalStateManager.maybePanelTabsControllerOf(
+    _sectionsController = GlobalStateManager.maybeSectionsControllerOf(
       context,
     )!;
     if (_orderedSections == null) {
@@ -107,7 +107,7 @@ class _AddonControlSectionListState extends State<_AddonControlSectionList> {
     setState(() {
       _orderedSections!.move(oldIndex, fixedNewIndex);
     });
-    _panelTabsController.reorder(
+    _sectionsController.reorder(
       tab: widget.tab,
       newSectionsOrder: _orderedSections!,
       getId: (s) => s.id,
@@ -117,10 +117,10 @@ class _AddonControlSectionListState extends State<_AddonControlSectionList> {
   void onToggleVisibility(int index) {
     final section = _orderedSections![index];
     setState(() {
-      _panelTabsController.setVisibility(
+      _sectionsController.setVisibility(
         widget.tab,
         section.id,
-        visible: !_panelTabsController.isVisible(widget.tab, section.id),
+        visible: !_sectionsController.isVisible(widget.tab, section.id),
       );
     });
   }
@@ -133,7 +133,7 @@ class _AddonControlSectionListState extends State<_AddonControlSectionList> {
           ControlSection(
             id: section.id,
             title: section.addonControlSection.title,
-            visible: _panelTabsController.isVisible(widget.tab, section.id),
+            visible: _sectionsController.isVisible(widget.tab, section.id),
             children: section.addonControlSection.children,
           ),
       ],